Deduplicate chartmap IO and test roundtrip
Factor chartmap NetCDF writing into python/libneo/chartmap_io.py, route STL writer through shared IO, and share Fortran chartmap roundtrip checks via chartmap_test_utils (issue #183).
