+++ test/CMakeLists.txt
@@ -278,8 +278,11 @@ set_tests_properties(test_chartmap_validator PROPERTIES
   FAIL_REGULAR_EXPRESSION "STOP"
   WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR})
 
+add_library(chartmap_test_utils STATIC source/chartmap_test_utils.f90)
+target_link_libraries(chartmap_test_utils PUBLIC neo)
+
 add_executable(test_vmec_chartmap_generator.x source/test_vmec_chartmap_generator.f90)
-target_link_libraries(test_vmec_chartmap_generator.x neo)
+target_link_libraries(test_vmec_chartmap_generator.x chartmap_test_utils)
 add_test(NAME test_vmec_chartmap_generator
   COMMAND test_vmec_chartmap_generator.x)
 set_tests_properties(test_vmec_chartmap_generator PROPERTIES
@@ -288,7 +291,7 @@ set_tests_properties(test_vmec_chartmap_generator PROPERTIES
   DEPENDS setup_vmec_wout)
 
 add_executable(test_python_chartmap_generator.x source/test_python_chartmap_generator.f90)
-target_link_libraries(test_python_chartmap_generator.x neo)
+target_link_libraries(test_python_chartmap_generator.x chartmap_test_utils)
 add_test(NAME test_python_chartmap_generator
   COMMAND test_python_chartmap_generator.x)
 set_tests_properties(test_python_chartmap_generator PROPERTIES
